move to core.js
